% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-mf.R
\name{fuzzy_partitions}
\alias{fuzzy_partitions}
\title{Define the fuzzy parititions}
\usage{
fuzzy_partitions(...)
}
\arguments{
\item{...}{are the diferent partitions, e.g., \code{trapezoid_mf} or \code{triangle_mf}.}
}
\value{
the \code{(fuzzy_partitions <- list(...)}
}
\description{
It is a constructor of fuzzy partitions, it defines a set of membership functions.  It takes as input a set of
\code{trapezoid_mf} or \code{triangle_mf} or objects in the shape_mf class.
}
\examples{
fuzzy_partitions(triangle_mf(450,450,550),
                 triangle_mf(450,550,600),
                 trapezoid_mf(550,600,800, 1000),
                 triangle_mf(800,1000,1300),
                 trapezoid_mf(1000,1300,1500,1500))
}

