% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-rules.R
\name{fuzzy_rules}
\alias{fuzzy_rules}
\title{Define the fuzzy rules}
\usage{
fuzzy_rules(...)
}
\arguments{
\item{...}{one or more \code{\link{fuzzy_rule}} objects.}
}
\value{
\code{fuzzy_rules <- list(...)}
}
\description{
It is a constructor of fuzzy rules, the arguments are the diferent \code{\link{fuzzy_rule}} object.
}
\examples{

fuzzy_rules(fuzzy_rule(0,0,1, 0,0,1, 0,0,1),
     fuzzy_rule(1,0,0, 1,0,0, 1,0,0),
     fuzzy_rule(0,1,0, 0,1,0, 0,1,0))
}

