% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-ldcp.R
\name{generate_code}
\alias{generate_code}
\title{Generate the R code}
\usage{
generate_code(input, output)
}
\arguments{
\item{input}{is the XML source path file. E.g. "/folder/ldcp_system.xml".}

\item{output}{is the R destination path file. E.g. "/folder/ldcp_system.R".}
}
\value{
If the process ends without error, the user will receive a message that indicates
 that the code has been generated successfully. Otherwise, the user will receive the
 detailed list of errors.
}
\description{
The function takes as input the path to a XML file that contains a LDCP system.
Then it generates its corresponding in R code. This R code is stored in an output
file. The output file path is another function parameter.
}
\examples{
\dontrun{generate_code('extdata/comfortableroom','comfortableroom')}

## The code has been generated successfully
}
