% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-rules.R
\name{operator}
\alias{operator}
\title{Define the operator}
\usage{
operator(conjunction, disjunction)
}
\arguments{
\item{conjunction}{is the method used to make the conjunction.}

\item{disjunction}{is the method used to make the disjunction.}
}
\value{
the \code{opertator} object \code{my_op <- list(conjunction, disjunction)}.
}
\description{
The operator defines the conjunction and disjunction functions used in the fuzzy rules.
It takes as input parameters the function used to implement the conjunction,
and the function used to implement the disjunction, e.g., "operator(min, max)", where
min and max are functions defined by the R language that calculate the maximum and
minimum, respectively, from a set of values received as input. Note that, we implicitly assign
to the fuzzy implication operator (THEN) the function given for conjunction
}
\examples{
operator <- operator(min, max)
}

