% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{pm_multidimensional}
\alias{pm_multidimensional}
\title{Define the pm of a multidimensional cp}
\usage{
pm_multidimensional(...)
}
\arguments{
\item{...}{the set of \code{\link{pm}}s}
}
\value{
The generated \code{pm_multidimensional <- list(...)}
}
\description{
It is a set of \code{\link{pm}}s that infer a multidimensional \code{\link{cp}}.
}
