% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-mf.R
\name{trapezoid_mf}
\alias{trapezoid_mf}
\title{Define the trapezoid membership function}
\usage{
trapezoid_mf(a, b, c, d)
}
\arguments{
\item{a}{the trapezoid point a.}

\item{b}{the trapezoid point b.}

\item{c}{the trapezoid point c.}

\item{d}{the trapezoid point d.}
}
\value{
the \code{(trapezoid_mf <- list(a,b,c,d))}
}
\description{
It is a constructor of trapezoidal shapes. They take as input the numerical values which
define the anchor points in membership functions.
}
\examples{
trapezoid_mf(0, 1, 2, 3)
}

