% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy-mf.R
\name{triangle_mf}
\alias{triangle_mf}
\title{Define the triangle membership function}
\usage{
triangle_mf(a, b, c)
}
\arguments{
\item{a}{the trapezoid point a.}

\item{b}{the trapezoid point b.}

\item{c}{the trapezoid point c.}
}
\value{
the \code{(triangle_mf <- list(a,b,c))}
}
\description{
It is a constructor of triangular shapes. They take as input the numerical values which
define the anchor points in membership functions.
}
\examples{
triangle_mf(0, 1, 2)
}
