% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-ldcp.R
\name{validate_xml}
\alias{validate_xml}
\title{Validate the XML file}
\usage{
validate_xml(xmlfile, schema = NULL)
}
\arguments{
\item{xmlfile}{is the XML source path file. E.g. "/folder/ldcp_system.xml".}

\item{schema}{is the ldcp schema path file. By default is "ldcpSchema.xsd".}
}
\value{
If the process ends without error, the user will receive the message that indicates
that the XML is valid. Otherwise, the user will receive the detailed list of errors.
}
\description{
The function takes as input the path to a XML file that contains a LDCP system.
Then it validates the LDCP system.
}
\examples{
\dontrun{validate_xml('extdata/comfortableroom.xml')}

## The xml is valid

}

