% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml-ldcp.R
\name{xml2rldcp}
\alias{xml2rldcp}
\title{XML to rLDCP}
\usage{
xml2rldcp(input, output)
}
\arguments{
\item{input}{is the XML source path file. E.g. "/folder/ldcp_system.xml".}

\item{output}{is the R destination path file. E.g. "/folder/ldcp_system.R".}
}
\value{
If the process ends without error, the user will receive two messages:
one indicates that the XML is valid and the other indicates that the code has
been generated successfully. Otherwise, the user will receive the detailed list of errors.
}
\description{
The function takes as input the path to a XML file that contains a LDCP system.
Then it validates the LDCP system and generates its corresponding in R code.
This R code is stored in an output file. The output file path is another
function parameter.
}
\examples{
\dontrun{xml2rldcp('extdata/comfortableroom.xml','comfortableroom.R')}

## The xml is valid
## The code has been generated successfully
}
